Datacentric: temporary fix for page faults
Temporarily, not to sample every page-faults, but instead sample for a
certain frequency. We should use mprotect instead of perf_events'
page-fault to enable mask and unmask the memory protection. However,
this requires careful handle of sigsegv. (and we want to avoid tightly
couple datacentric and hpcrun since datacentric is a "plugin".


Minor fix:
- hpcprof: outpur the period mean as float
- fix hpcfnbounds to only send static variables that match datacentric
minimum bytes requirement. (sharing code between hpcrun-datacentric and
hpcfnbounds in support-lean).
- perf-util: make perf_event to wake up for every overflows.

+++ src/lib/prof/CallPath-Profile.cpp
@@ -761,10 +761,10 @@ Profile::writeXML_hdr(std::ostream& os, uint metricBeg, uint metricEnd,
     // Metric
     os << "    <Metric i" << MakeAttrNum(i)
        << " n" << MakeAttrStr(m->name())
-       << " v=\"" << m->toValueTyStringXML() << "\""
-       << " em=\"" << m->isMultiplexed()    << "\""
-       << " es=\"" << m->num_samples()      << "\""
-       << " ep=\"" << long(m->periodMean())       << "\""
+       << " v=\""  << m->toValueTyStringXML() << "\""
+       << " em=\"" << m->isMultiplexed()      << "\""
+       << " es=\"" << m->num_samples()        << "\""
+       << " ep=\"" << m->periodMean()         << "\""
        << " t=\"" << Prof::Metric::ADesc::ADescTyToXMLString(m->type()) << "\"";
     if (m->partner()) {
       os << " partner" << MakeAttrNum(m->partner()->id());
